src/test/java/org/kakao23/P5_3_Merge_TableTest.java
package org.kakao23;

import org.assertj.core.api.Assertions;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class P5_3_Merge_TableTest {

    @Test
    void solution() {
        // given
        String[] commands = {"UPDATE 1 1 menu",
                "UPDATE 1 2 category",
                "UPDATE 2 1 bibimbap",
                "UPDATE 2 2 korean",
                "UPDATE 2 3 rice",
                "UPDATE 3 1 ramyeon",
                "UPDATE 3 2 korean",
                "UPDATE 3 3 noodle",
                "UPDATE 3 4 instant",
                "UPDATE 4 1 pasta",
                "UPDATE 4 2 italian",
                "UPDATE 4 3 noodle",
                "MERGE 1 2 1 3",
                "MERGE 1 3 1 4",
                "UPDATE korean hansik",
                "UPDATE 1 3 group",
                "UNMERGE 1 4",
                "PRINT 1 3",
                "PRINT 1 4"};
        String[] expected = {"EMPTY", "group"};

        // when
        P5_3_Merge_Table p5 = new P5_3_Merge_Table();
        String[] answer = p5.solution(commands);

        // then
        Assertions.assertThat(answer).isEqualTo(expected);
    }

    @Test
    void solution2() {
        // given
        String[] commands = {"UPDATE 1 1 a",
                "UPDATE 1 2 b",
                "UPDATE 2 1 c",
                "UPDATE 2 2 d",
                "MERGE 1 1 1 2",
                "MERGE 2 2 2 1",
                "MERGE 2 1 1 1",
                "PRINT 1 1",
                "UNMERGE 2 2",
                "PRINT 1 1"};
        String[] expected = {"d", "EMPTY"};

        // when
        P5_3_Merge_Table p5 = new P5_3_Merge_Table();
        String[] answer = p5.solution(commands);

        // then
        Assertions.assertThat(answer).isEqualTo(expected);
    }


}
